Create generate current file
Relates #27
