+++ utils/submitEventListener.js
@@ -2,11 +2,12 @@ import {
   showBufferingOverlay,
   hideBufferingOverlay,
 } from "./helpers/bufferingOverlay.js";
-import { createElement, appendElements } from "./helpers/handleElements.js";
+import { createElement } from "./helpers/handleElements.js";
 import { getMessage } from "./api/getMessage.js";
 import { handleApiRequests } from "./handleApiRequests.js";
 import { generateForecast } from "./generateForecast.js";
 import { generateSun } from "./generateSun.js";
+import { generateCurrent } from "./generateCurrent.js";
 
 export const submitEventListener = () => {
   const form = document.querySelector("form");
@@ -82,33 +83,13 @@ export const submitEventListener = () => {
 
         image = newImage;
         mapDiv.appendChild(newImage);
-        //current
-        const temperatureHeading = createElement("h3", {
-          textContent: `${current_temp}°`,
-        });
-        const windSpeedHeading = createElement("h3", {
-          textContent: `${current_windSpeed} km/h`,
-        });
 
-        const current_icon = createElement("img", {
-          src: `./utils/icons/${current_temp_code}.svg`,
-          className: "icon",
-          alt: "forecast icon",
-        });
-
-        const wind_icon = createElement("img", {
-          src: `./utils/icons/wind.svg`,
-          className: "icon",
-          alt: "Wind speed con",
-        });
-        current.innerHTML = "";
-        appendElements(current, [
-          current_icon,
-          temperatureHeading,
-          wind_icon,
-          windSpeedHeading,
-        ]);
-
+        generateCurrent(
+          current_temp,
+          current_windSpeed,
+          current,
+          current_temp_code
+        );
         generateSun(s_rise, s_set, currentsun, currSunrise, currSunset);
         generateForecast(date_arr, fore_code, fore_max, fore_min);
         hideBufferingOverlay();
